requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
choreographer==1.2.1
contourpy==1.3.3
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
docopt-ng==0.9.0
et_xmlfile==2.0.0
fonttools==4.61.1
ftpretty==0.4.0
goatools==1.5.2
idna==3.11
iniconfig==2.3.0
joblib==1.5.3
kaleido==1.2.0
kiwisolver==1.4.9
logistro==2.0.1
markdown-it-py==4.0.0
matplotlib==3.10.8
mdurl==0.1.2
narwhals==2.15.0
numpy==2.4.2
openpyxl==3.1.5
orjson==3.11.6
packaging==26.0
pandas==3.0.0
patsy==1.0.2
pillow==12.1.0
plotly==6.5.2
plotly-express==0.4.1
pluggy==1.6.0
pydot==4.0.1
Pygments==2.19.2
pyparsing==3.3.2
pytest==9.0.2
pytest-timeout==2.4.0
python-dateutil==2.9.0.post0
requests==2.32.5
rich==14.3.1
scikit-learn==1.8.0
scipy==1.17.0
setuptools==80.10.2
simplejson==3.20.2
six==1.17.0
statsmodels==0.14.6
threadpoolctl==3.6.0
urllib3==2.6.3
xlsxwriter==3.2.9
